requirements.txt
gradio>=5.0.0
openai>=1.0.0
pillow>=10.0.0
pydantic>=2.0.0
